refactoring: use "aiosqlite" (#309)
Initially, we wrote separate logic for the _Server_ mode and the home
use mode (_DEFUALT_), since PGSQL was used for the server, which
supports asynchronous calls.

This led to terrible code duplication, and to the fact that during
development we use SQLite more and many bugs and typos for the server
mode remain unnoticed for a long time.

This PR adds the use of the **aiosqlite** wrapper library - which will
allow us to remove unnecessary code and, if possible, unify the logic.

---------

Signed-off-by: Alexander Piskun <bigcat88@icloud.com>