Feature: Surprise Me (#314)
Often you can run out of imagination for prompts, and it would be nice
for LLM to compose them for you - to get inspiration.

This PR adds a `Surprise Me` button to flows when LLM generates a prompt
for you - completely random or on a given topic.

The topic is set in the prompt, if it is not empty, then it is used.

---------

Signed-off-by: Andrey Borysenko <andrey18106x@gmail.com>
Co-authored-by: Andrey Borysenko <andrey18106x@gmail.com>