feat: big rework of base models directory (for docker container support) (#324)
These are the main changes from backend for the upcoming version
**1.12**

1. The endpoint `/settings/comfyui/folders/autoconfig` is removed -
instead of it, UI should simply set the root directory setting for
models from the UI in the same place (**from the admin's point of view,
nothing has changed**). UI should ask to perform the restart of
Visionatrix after setting this value to changes take effect.
2. Visionatrix always checks for the presence of `comfyui_models_folder`
setting in the database when starting, if it is not there - it asks for
its value in the console and sets it.
3. Support for the `extra_model_paths.yaml` file in the **Visionatrix
directory** has been removed (support for the file inside the ComfyUI
directory remains)
4. Added console commands: `list-global-settings`, `get-global-setting`,
`set-global-setting`

## Why was this needed

This PR changes the way we run ComfyUI, and sets `models_dir` inside
ComfyUI which is used by most custom nodes to download their models.

From what I tested, now all models are ALWAYS downloaded to the
specified directory, **which will finally allow us to start releasing
docker images** with version `1.13` that will not break every version,
because of custom nodes from the community, which download models
"wherever they want".

---------

Signed-off-by: bigcat88 <bigcat88@icloud.com>
Signed-off-by: Andrey Borysenko <andrey18106x@gmail.com>
Co-authored-by: Andrey Borysenko <andrey18106x@gmail.com>