fix: deploying on Novita.a/RunPodi;optimized Dockerfile (#354)
Today, when I was testing the docker image deployment, I realized that
it doesn't work on [Novita.ai](https://novita.ai) - because we have
`VOLUME ["/data"]` defined in our Dockerfile - which causes an error, if
do not setup such volume manually(which will be strange to do).

This should be fixed with this PR.

Also added `DOCKER_BUILDKIT=1` to greatly speed up building of docker
images by using docker cache.

Signed-off-by: bigcat88 <bigcat88@icloud.com>