fix: invalid model download path for FrameInterpolation node (#357)
In version `2`, when we completely switched to ComfyUI `base_path` and
`models_path`, we forgot that the FrameInterpolation node always stores
models in a local folder.

Just bring back the check, if the file name starts with `custom_nodes`,
specify this relative path to the ComfyUI root directory, not
`base_path`.

Signed-off-by: bigcat88 <bigcat88@icloud.com>