feat(settings): add cache_type, cache_size settings for Workers (#404)
Related to #400

In this PR we remove the command line options of ComfiUI related to
configuring caches, and add their configuration via UI.

By default, the behaviour does not change, the `classic` cache will be
used.

Again, it is necessary for docker installations, and it is more
convenient to change these settings without rebooting.

Note: support for command line arguments `--cache-classic`,
`--cache-lru` was dropped as this now should be done through UI or by
setting values in the database(`python3 -m visionatrix
set-global-setting --key=cache_type --value=lru`).

---------

Signed-off-by: Andrey Borysenko <andrey18106x@gmail.com>
Signed-off-by: bigcat88 <bigcat88@icloud.com>
Co-authored-by: bigcat88 <bigcat88@icloud.com>