fix(UI): Reset Params button: detect default value of the Aspect Ratio parameter (#437)
Previous incorrect behavior could be seen in the `Animal Clothing` or
`Juggernaut Lite` flows, where the default "Aspect Ratio" is "2:3"
rather than "1:1" which the UI resets the settings to.

Also, added the mapping to the backend of the new `20:11 (1280x704)`
`Aspect Ratio` for the new upcoming **Cosmos Predict 2** flow where such
aspect ratio is the default one.

Signed-off-by: bigcat88 <bigcat88@icloud.com>